move reference-logic into refs-folder, add calls-folder for call hierarchy work
