+++ src/refs/editorHighlights.ts
